chromium: import patches from alpine
The deadlock patch is most likely not limited to musl but its a race
condition and it seems like with the current way how it works its only
triggered with musl and with namespaces enabled.

The other patch just makes pthreads in v8 use CLOCK_MONOTONIC like it
does for glibc.